fix(RawImageDisplayArea): add ack_id echo and user_data event when frame received
When a frame is sent by the server to the RawImageDisplayArea, allow
the inclusion of `ack_id` in the meta information, which is sent
as message back to the server when the frame is received. Allows the
server to track if the client is keeping up with the frames sent.

Also allow `user_data` in the meta information, which is propagated
to the client as a Vue event.